fix for issue #968

+++ publish/scripts/installer.js
@@ -798,7 +798,7 @@ module.exports = function($logger, $projectData, hookArgs) {
         var platform = hookArgs['checkForChangesOpts']['platform'].toLowerCase(); // ios | android
         var platformsDir = hookArgs['checkForChangesOpts']['projectData']['platformsDir'];
         var appResourcesDirectoryPath = hookArgs['checkForChangesOpts']['projectData']['appResourcesDirectoryPath'];
-        var forcePrepare = true; // whether to force NS to run prepare
+        var forcePrepare = true; // whether to force NS to run prepare, defaults to true
         var npfInfoPath = path.join(platformsDir, platform, ".pluginfirebaseinfo");
         var nsPrepareInfoPath = path.join(platformsDir, platform, ".nsprepareinfo");
         var copyPlistOpts = { platform, appResourcesDirectoryPath, buildType, $logger }
@@ -818,10 +818,10 @@ module.exports = function($logger, $projectData, hookArgs) {
             $logger.info('nativescript-plugin-firebase: running release build or change in environment detected, forcing prepare!');
 
             if (fs.existsSync(npfInfoPath)) { fs.unlinkSync(npfInfoPath); }
-            fs.unlinkSync(nsPrepareInfoPath);
+            if (fs.existsSync(nsPrepareInfoPath)) { fs.unlinkSync(nsPrepareInfoPath); }
 
             if (copyPlist(copyPlistOpts)) { resolve(); } else { reject(); }
-        } else { if (copyPlist(copyPlistOpts)) { resolve(); } else { reject(); } }
+        } else { resolve(); }
     });
 };
